requirements.txt
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth==1.6.3
django-googlecalendar==3.0.0
requests==2.22.0
